+++ attriax/lib/src/internal/attriax_session_lifecycle_manager.dart
@@ -335,7 +335,7 @@ class AttriaxSessionLifecycleManager with WidgetsBindingObserver {
 
     await _requestManager.enqueue(
       attriaxBuildTrackSessionRequest(
-        appToken: _config.projectToken,
+        projectToken: _config.projectToken,
         deviceIdSource: _sessionManager.requireDeviceIdSource(),
         session: session,
         kind: kind,